add `plot_distribution` (#59)
* create graphics.py, move plot_bitinformation

* Update graphics.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update graphics.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update environment.yml

* Update requirements.txt

* Update __init__.py

* Update graphics.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update graphics.py

* Update graphics.py

* Update graphics.py

* Update graphics.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update graphics.py

* Update graphics.py

* Update bitinformation_pipeline.py

* Update bitinformation_pipeline.py

* add to quick-start as comment

* Update quick-start.ipynb

* Update quick-start.ipynb

* Update quick-start.ipynb

* Update graphics.py

* Update requirements.txt

* Update environment.yml

* Update graphics.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
