+++ bitinformation_pipeline/__init__.py
@@ -5,7 +5,7 @@ from .bitinformation_pipeline import (
     get_bitinformation,
     get_keepbits,
     get_prefect_flow,
-    plot_bitinformation,
 )
 from .bitround import jl_bitround, xr_bitround
+from .graphics import plot_bitinformation, plot_distribution
 from .save_compressed import get_compress_encoding_nc, get_compress_encoding_zarr
